refactor: replace keytar package with sfdx keychain implementation
The npm package `keytar` depends on Python which we want to get rid of.
`@salesforce/core` provides a keychain implementation for Windows/MacOS/Unix.

[HUT-245]

BREAKING CHANGE: For existing installations you might be prompted once
to grant access to the hutte-io credential in your keychain of your OS.

For example on MacOS, the process accessing the keychain was changed from "node" to "security".
So you might be prompted:

> security wants to use your confidential information stored in "hutte-io" in your keychain.
>
> To allow this, enter the "login" keychain password.

Here you can enter your keychain password and click "Always Allow".