feat: change storage implementation for API token
BREAKING CHANGE: We no longer use the deprecated 'keytar' library to store the API token.
From now on we use the the same library as the Salesforce CLI.

The 'keytar' dependency will be removed in an upcoming major version.
When using 'hutte' the next time, the API token will be stored encrypted in a new config file.
You can then delete the 'hutte-io' entry from the keychain of your OS.

MacOS

- Open 'Keyhain Access.app'
- Search for 'hutte-io' of kind application password
- Right click and delete

Linux

- TBD

Windows

- TBD